+++ src/main/java/org/noses/arnoldshortenator/ArnoldShortenatorApplication.java
@@ -1,4 +1,4 @@
-package org.noses.urlshortener;
+package org.noses.arnoldshortenator;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
