+++ src/test/java/org/noses/arnoldshortenator/ArnoldShortenatorApplicationTests.java
@@ -1,4 +1,4 @@
-package org.noses.urlshortener;
+package org.noses.arnoldshortenator;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
